+++ Code/ModUtil.cs
@@ -1,6 +1,11 @@
 ﻿using Mono.Cecil.Cil;
 using MonoDetour.Cil;
 using MonoMod.Cil;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using R2API;
+using RoR2;
+using SimpleJSON;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -9,12 +14,6 @@ namespace LordsItemEdits
 {
     internal static class ModUtil
     {
-        internal static string GetLangFileLocation(string fileName)
-        {
-            string langFolderPath = System.IO.Path.Combine(System.IO.Path.GetDirectoryName(Plugin.PluginInfo.Location), "Language");
-            return System.IO.Path.Combine(langFolderPath, (fileName + ".language"));
-        }
-
         internal static void LogILError(ILContext il, ILCursor c)
         {
             Log.Error($"COULD NOT IL HOOK {il.Method.Name}");